Created Provider and presenter.js in Timer
