+++ App.js
@@ -1,9 +1,17 @@
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import Timer from './components/Timer'
+import reducer from './reducer';
+import { createStore } from 'redux';
+import { Provider } from 'react-redux';
 
+let store = createStore(reducer);
 export default class App extends React.Component {
   render() {
-    return <Timer></Timer>;
+    return (
+      <Provider store={store}>
+        <Timer></Timer>
+      </Provider>
+    );
   }
 }
